requirements.txt
aircv==1.4.6
altgraph==0.17.2
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.10.0
et-xmlfile==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.3.1
line-bot-sdk==1.18.0
MarkupSafe==2.1.1

MouseInfo==0.1.3


pefile==2022.5.30
Pillow==8.3.0
psycopg2==2.9.3

pycparser==2.21
PyGetWindow==0.0.9

pyinstaller-hooks-contrib==2022.7
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.1
PyTweening==1.0.3
six==1.16.0
SQLAlchemy==1.4.40
urllib3==1.26.6
Werkzeug==2.2.2
zipp==3.8.1
